requirements.txt
pillow
cython
matplotlib
tweepy
requests
praw